Remove secp256kfun for k256
It should be the more established library. It also lets us remove a 
backend file for a few lines of configuration.

Also adds in p256 which is effectively secp256kfun with a single letter 
changed (k256 -> p256).

Adds a missing Ristretto signature test.

Ideologically rotates the library by aligning curves by dalek, ff/group. 
This prioritizes dalek as what support should be based around instead of 
the generally more modern ff/group curves. This was done to ensure 
widespread compatibility and to focus on what's stable and required 
versus what's new and available.