Remove PlasmaPower as an author
He is THE author of this proof and the only reason this library exists. 
That said, I have decided to move it out of the project we built 
together into its own library. As part of that, I am signing up to 
maintain it and advance it, and I do not wish him to be contacted and 
bothered over my work. LICENSE still properly includes him, not just 
fulfilling a legal obligation, yet giving him the recognition he 
rightfully deserves.